Adds a basic component to display ticker info on the top right of the GUI
