+++ web/src/app/pages/donate/donate.component.scss
@@ -3,7 +3,7 @@
 }
 
 .full-card {
-    width: 30%;
+    width: 50%;
     margin: 0px auto !important;
 }
