+++ web/src/app/shared/all-updates-ticker/all-updates-ticker.component.html
@@ -1,0 +1,3 @@
+<div *ngIf="showTicker" class="one-time-animation" >
+  <p>{{message}}</p>
+</div>
