+++ web/src/app/shared/all-updates-ticker/all-updates-ticker.component.scss
@@ -1,5 +1,5 @@
 .one-time-animation {
-    animation: one-time-animation 2s forwards 1;
+    animation :  one-time-animation 2s forwards 1;
   }
   @keyframes one-time-animation {
     from {
@@ -16,4 +16,6 @@
     box-shadow: 0 0 10px green inset;
     transition: box-shadow 2s ease;
   }
-  
+  p {
+    font-size:14px;
+  }
