+++ web/src/app/shared/all-updates-ticker/all-updates-ticker.component.spec.ts
