--- web/src/app/shared/exchange-currency-ticker/exchange-currency-ticker.component.html
@@ -1,28 +1,0 @@
-<mat-card class="exchange-card one-time-animation"  >
-  <mat-card-header>
-      <mat-card-title>{{ticker?.Exchange}} {{ticker?.CurrencyPair}} Ticker update</mat-card-title>
-  </mat-card-header>
-  <mat-card-content>
-      <mat-grid-list cols="3" rowHeight="16:9" >
-          <mat-grid-tile>
-              <p>Last: {{ticker?.Last | number:'1.0-1'}}</p>
-            </mat-grid-tile>
-          <mat-grid-tile>
-              <p> Low: {{ticker?.Low | number:'1.0-1'}}</p>
-          </mat-grid-tile>
-          <mat-grid-tile>
-              <p> High: {{ticker?.High | number:'1.0-1'}}</p>
-          </mat-grid-tile>
-          
-          <mat-grid-tile>
-              <p> Bid: {{ticker?.Bid | number:'1.0-1'}}</p>
-            </mat-grid-tile>
-            <mat-grid-tile>
-                <p> Ask: {{ticker?.Ask | number:'1.0-1'}}</p>
-              </mat-grid-tile>
-              <mat-grid-tile>
-                  <p>Volume: {{ticker?.Volume | number:'1.0-1'}}</p>
-                </mat-grid-tile>
-      </mat-grid-list>
-  </mat-card-content>
-
