--- web/src/app/shared/exchange-currency-ticker/exchange-currency-ticker.component.spec.ts
@@ -1,25 +1,0 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { ExchangeCurrencyTickerComponent } from './exchange-currency-ticker.component';
-
-describe('ExchangeCurrencyTickerComponent', () => {
-  let component: ExchangeCurrencyTickerComponent;
-  let fixture: ComponentFixture<ExchangeCurrencyTickerComponent>;
-
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      declarations: [ ExchangeCurrencyTickerComponent ]
-    })
-    .compileComponents();
-  }));
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(ExchangeCurrencyTickerComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should be created', () => {
-    expect(component).toBeTruthy();
-  });
-});
