+++ web/src/app/shared/navbar/navbar.component.html
@@ -7,6 +7,7 @@
         </a>
         <app-selected-currency></app-selected-currency>
         <div class="flex-spacer"></div>
+        <app-all-updates-ticker></app-all-updates-ticker>
         <theme-picker></theme-picker>
         &nbsp;
         &nbsp;
